Comando de projeto por diretorio (.porecatu) (#67)
* feat: parser e confianca do arquivo de projeto .porecatu

Implementa a etapa 2 de 4 do PRD-012/ADR-0051: seção [project_file] em
Config, parser de seções cruas do .porecatu (primeira seção com nome
duplicado vence -- regra nova, documentada na referência), checagem de
confiança por allowlist canonicalizada e comparada por componente de
caminho, e resolve() com a ordem de garantias do PRD-012. [project_file]
classificado como recarga C em porecatu-ui. Exemplo descomentado em
porecatu.example.toml.


* feat: gatilho e escrita do .porecatu no PTY na restauracao de aba

Implementa a etapa 3 de 4 do PRD-012/ADR-0051, em porecatu-ui: distingue
aba restaurada de aba nova (SpawnOrigin) no ponto de spawn, resolve o
.porecatu so para aba Restored com cwd existente (RF-12.12/RF-12.14),
escreve o comando no PTY apos o primeiro byte de saida e um silencio de
300ms (com teto de 5s e guarda de tela alternativa, ADR-0051 SS6), e mostra
a nota de diretorio nao autorizado uma vez por execucao do app (RF-12.6).
Arquivo ilegivel vira nota por aba na hora (RF-12.4). A decisao de quando
escrever, a montagem dos bytes e as duas regras "uma vez" saem como
funcoes puras testaveis sem Instant::now()/Terminal real.

Corrige de passagem um cast redundante em input.rs que quebrava
`cargo clippy --all-targets` independente desta mudanca.


* feat: fecha o comando de projeto por diretorio (.porecatu) com verificacao ponta a ponta

Etapa 4 de 4 do PRD-012/ADR-0051: verificacao ponta a ponta no app de
verdade (sessao gravada a mao, powershell.exe nesta maquina) e fechamento
da documentacao.

Achou e corrigiu um bug real: a nota de ".porecatu nao autorizado" nao
contribuia nenhum Instant para schedule_next_wake, entao so aparecia por
coincidencia com outro temporizador do processo (o debounce de gravacao
de sessao) -- as vezes segundos depois do foco na aba, quando RF-12.6 nao
tem criterio de tempo nenhum. check_project_file_notice() sai de tick_all
(so roda em ResumeTimeReached) e passa a rodar em schedule_next_wake (toda
volta do event loop); confirmado ao vivo, a nota passou a aparecer em
menos de 500ms.

Cinco cenarios confirmados ao vivo: o caso motivador (diretorio
autorizado, comando escrito e executado no prompt), diretorio nao
autorizado (uma nota so, nada executado), aba nova nao dispara,
restauracao preguicosa (so a ativa roda no arranque, as outras no
primeiro foco) e o comando visivel/no historico/interrompivel por Ctrl+C.

Documentacao: removidos os tres avisos de "ainda nao implementado"
(guia do usuario, referencia do arquivo, CLAUDE.md), CLAUDE.md reescrito
para o estado real com uma entrada nova na secao de descobertas, roadmap
com o que de fato aconteceu em cada etapa e a divida de verificacao
(so Windows, so powershell/cmd), CHANGELOG cobrindo o recurso inteiro.

Suite: 706 testes, clippy -D warnings limpo, fmt --check limpo.


* docs: documentacao e decisao do arquivo de projeto .porecatu (etapa 1/4)

Etapa 1 de 4 do comando de projeto por diretorio, fora da ordem de fases:
PRD-012 (RF-12.1 a RF-12.15) e ADR-0051, que fecham as cinco decisoes que
o PRD deixa em aberto -- formato de secoes cruas, casamento exato de
shell_name sem tabela de apelidos, confianca por allowlist vazia por
default, quando escrever no PTY, e a nota para arquivo nao autorizado.

Acompanha: indice do README e da tabela de ADRs; PRD-003 ganha uma emenda
em blockquote deixando explicito que a proibicao de reexecutar processo
observado continua valendo -- o .porecatu executa o que o usuario
declarou, nao o que o app gravou; arquitetura.md SS6.1 separa "config do
app" de "arquivo do projeto" e onde cada metade do codigo mora.

Ficou sem commit proprio nas etapas 2 a 4 (parser/confianca em
porecatu-config, gatilho/escrita em porecatu-ui, verificacao ponta a
ponta) -- commitado agora, retroativamente, a pedido do usuario.


---------

+++ crates/porecatu-ui/src/input.rs
@@ -453,8 +453,8 @@ mod tests {
 
         for &scale in &[1.0_f32, 1.25, 1.5, 2.0] {
             let physical_cursor = (
-                logical_cursor.0 as f64 * scale as f64,
-                logical_cursor.1 as f64 * scale as f64,
+                logical_cursor.0 * scale as f64,
+                logical_cursor.1 * scale as f64,
             );
             let logical_from_physical = (
                 (physical_cursor.0 / scale as f64) as f32,
